# docs/specs/openapi-snapshot.json
{
  "components": {
    "schemas": {
...
      "CollectJobInfo": {
        "description": "一条已经排在 APScheduler 里的采集 job。",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "next_run_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Run Time"
          },
          "platform": {
            "title": "Platform",
            "type": "string"
          }
        },
        "required": [
          "id",
          "platform"
        ],
        "title": "CollectJobInfo",
        "type": "object"
      },
      "ConfigUpdateResponse": {
        "properties": {
          "changed_fields": {
            "items": {
              "type": "string"
            },
            "title": "Changed Fields",
            "type": "array"
          },
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "platform": {
            "title": "Platform",
            "type": "string"
          },
          "requires_restart": {
            "default": false,
            "title": "Requires Restart",
            "type": "boolean"
          }
        },
        "required": [
          "platform",
          "config"
        ],
        "title": "ConfigUpdateResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlatformControlStatus": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "shadowed_by_env": {
            "items": {
              "type": "string"
            },
            "title": "Shadowed By Env",
            "type": "array"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "PlatformControlStatus",
        "type": "object"
      },
      "PlatformControlUpdate": {
        "additionalProperties": false,
        "description": "PUT 的请求体。只有一个字段，所以不搞 `exclude_unset` 那套\"没写的保持原值\" ——\n那套是为了一张表里多个字段准备的，这里只会让人以为可以省略。",
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "PlatformControlUpdate",
        "type": "object"
      },
      "PlatformControlUpdateResponse": {
        "properties": {
          "changed_fields": {
            "items": {
              "type": "string"
            },
            "title": "Changed Fields",
            "type": "array"
          },
          "platform_control": {
            "$ref": "#/components/schemas/PlatformControlStatus"
          },
          "requires_restart": {
            "default": false,
            "title": "Requires Restart",
            "type": "boolean"
          }
        },
        "required": [
          "platform_control"
        ],
        "title": "PlatformControlUpdateResponse",
        "type": "object"
      },
...
      "ScheduleStatus": {
        "description": "`GET /api/schedule`：配置 + 运行态两份**并排**。\n\n两份分开是必须的：只看配置会说谎（改了没生效），只看 job 也会说谎\n（`scheduler.enabled: false` 时一条 job 都没有，而配置里明明写着 08:00）。",
        "properties": {
          "collect_cron": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collect Cron"
          },
          "collect_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collect Limit"
          },
          "collect_platforms": {
            "items": {
              "type": "string"
            },
            "title": "Collect Platforms",
            "type": "array"
          },
          "effective_platforms": {
            "items": {
              "type": "string"
            },
            "title": "Effective Platforms",
            "type": "array"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/CollectJobInfo"
            },
            "title": "Jobs",
            "type": "array"
          },
          "master_enabled": {
            "default": true,
            "title": "Master Enabled",
            "type": "boolean"
          },
          "scheduler_enabled": {
            "title": "Scheduler Enabled",
            "type": "boolean"
          },
          "scheduler_running": {
            "default": false,
            "title": "Scheduler Running",
            "type": "boolean"
          },
          "shadowed_by_env": {
            "items": {
              "type": "string"
            },
            "title": "Shadowed By Env",
            "type": "array"
          },
          "skipped_platforms": {
            "items": {
              "type": "string"
            },
            "title": "Skipped Platforms",
            "type": "array"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "scheduler_enabled",
          "timezone",
          "collect_cron",
          "collect_limit"
        ],
        "title": "ScheduleStatus",
        "type": "object"
      },
      "ScheduleUpdate": {
        "description": "PUT 的请求体：**没出现的键保持原值**。\n\n所以 `collect_cron: null` 是\"关掉定时\"，而整个键不写是\"这次不改它\"。这两件事必须分得开，\n否则前端只能\"总是提交全表\"，那等于每次保存都把别人的字段顺手写一遍。",
        "properties": {
          "collect_cron": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collect Cron"
          },
          "collect_limit": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collect Limit"
          },
          "collect_platforms": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collect Platforms"
          }
        },
        "title": "ScheduleUpdate",
        "type": "object"
      },
      "ScheduleUpdateResponse": {
        "properties": {
          "changed_fields": {
            "items": {
              "type": "string"
            },
            "title": "Changed Fields",
            "type": "array"
          },
          "requires_restart": {
            "default": false,
            "title": "Requires Restart",
            "type": "boolean"
          },
          "schedule": {
            "$ref": "#/components/schemas/ScheduleStatus"
          }
        },
        "required": [
          "schedule"
        ],
        "title": "ScheduleUpdateResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Intelligence Hub V2",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/platform-control": {
      "get": {
        "description": "四家共用的那一个总闸。",
        "operationId": "get_platform_control_api_platform_control_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformControlStatus"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Platform Control",
        "tags": [
          "config"
        ]
      },
      "put": {
        "description": "写 `app.yaml` 的 `platform_control` 段 → 当场生效 → 重排采集 job → 发事件。\n\n与两条平台配置 PUT 同一条纪律：**运行时不听文件变化**，所以\"生效\"必须显式做。\n这里比它们少一步：没有需要推的快照 —— 注册表拿的是总闸的**闭包**、\n调度器与 runner 握着的是同一个 `AppConfig` 实例，而 `write_platform_control`\n原地换的就是那个实例上的字段（换指针会让没跟着换的持有者继续用旧值，\n症状是\"设置页说总闸已开，任务列表还是空的\"）。\n\n重排采集 job 必须在写盘**之后**：`reschedule_collect_jobs` 读的就是那份配置。",
        "operationId": "update_platform_control_api_platform_control_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlatformControlUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformControlUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Platform Control",
        "tags": [
          "config"
        ]
      }
    },
...
    "/api/platforms/{platform}/config": {
      "get": {
        "description": "平台配置（权威源来自 ConfigManager）+ 运行态健康镜像。\n\n健康取不到就是 `None`（从没检查过），不编一个 ok —— `PlatformRecord.is_healthy`\n只认显式 `\"ok\"`（V1 §7.20 的同一口径）。",
        "operationId": "get_platform_config_api_platforms__platform__config_get",
        "parameters": [
          {
            "in": "path",
            "name": "platform",
            "required": true,
            "schema": {
              "title": "Platform",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Platform Config Api Platforms  Platform  Config Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Platform Config",
        "tags": [
          "config"
        ]
      },
      "put": {
        "description": "校验 → 原子写盘 → 热重载 → 广播 `CONFIG_CHANGED` → 让适配器实例作废。",
        "operationId": "update_platform_config_api_platforms__platform__config_put",
        "parameters": [
          {
            "in": "path",
            "name": "platform",
            "required": true,
            "schema": {
              "title": "Platform",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Platform Config",
        "tags": [
          "config"
        ]
      }
    },
    "/api/platforms/{platform}/schema": {
      "get": {
        "description": "配置类的 JSON Schema，前端据此自动渲染表单（`ui:advanced` 折叠高级字段）。",
        "operationId": "platform_schema_api_platforms__platform__schema_get",
        "parameters": [
          {
            "in": "path",
            "name": "platform",
            "required": true,
            "schema": {
              "title": "Platform",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Platform Schema Api Platforms  Platform  Schema Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Platform Schema",
        "tags": [
          "config"
        ]
      }
    },
...
    "/api/schedule": {
      "get": {
        "description": "定时采集的当前配置，与已经排上的 job。",
        "operationId": "get_schedule_api_schedule_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleStatus"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Schedule",
        "tags": [
          "schedule"
        ]
      },
      "put": {
        "description": "写 `app.yaml` 的 scheduler 段 → 热更内存 → 当场重排采集 job → 发 `CONFIG_CHANGED`。\n\n与平台配置那条 PUT 同一条纪律：**运行时不听文件变化**，改了不生效是设计而不是 bug。\n所以\"生效\"这一步必须显式做，不许留给下一次启动。",
        "operationId": "update_schedule_api_schedule_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Schedule",
        "tags": [
          "schedule"
        ]
      }
    },
...
  }
}